Save Drive uploads to a "Screenity Videos" folder
Checks for a "Screenity Videos" folder, and creates one if it doesn't exist. Proceeds to then save all future uploads to that location.

Known issue: Screenity only has permissions to "see" folders that Screenity had previously created, so if a "Screenity Videos" folder was created by the user, Screenity will not see it and will create a new folder by that name (and then upload consistently to the folder created by Screenity.  This can also cause duplicate folder creation if the permission settings are refreshed (eg, if user deletes and then re-enables the extension), as that will cause Screenity to lose access to "see" folders created prior to the permission reset.